pyproject.toml
[tool.poetry]
name = "telegram-selenium-tools"
version = "0.1.0"
description = ""
authors = ["mihkevich <mihkevich.mm@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.17.2"
selenium-wire = "^5.1.0"
fake-useragent = "^1.4.0"
pillow = "^10.2.0"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
aiogram = "^3.4.1"
schedule = "^1.2.1"
requests = "^2.31.0"
pytest = "^8.1.1"
pytest-mock = "^3.14.0"
pytest-sugar = "^1.0.0"
mypy = "^1.9.0"
types-requests = "^2.31.0.20240406"
types-pillow = "^10.2.0.20240406"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
